Moved localization helper method to separate file
