+++ EssentialApp/EssentialAppTests/Helpers/ImageCommentsUIIntegrationTests+Localization.swift
@@ -1,0 +1,22 @@
+//
+//  ImageCommentsUIIntegrationTests+Localozation.swift
+//  EssentialAppTests
+//
+//  Created by Cronay on 24.12.20.
+//  Copyright © 2020 Essential Developer. All rights reserved.
+//
+
+import XCTest
+import EssentialFeed
+
+extension ImageCommentsUIIntegrationTests {
+	func localized(_ key: String, file: StaticString = #filePath, line: UInt = #line) -> String {
+		let table = "ImageComments"
+		let bundle = Bundle(for: ImageCommentsPresenter.self)
+		let value = bundle.localizedString(forKey: key, value: nil, table: table)
+		if value == key {
+			XCTFail("Missing localized string for key: \(key) in table: \(table)", file: file, line: line)
+		}
+		return value
+	}
+}
