+++ EssentialApp/EssentialAppTests/ImageCommentsUIIntegrationTests.swift
@@ -115,16 +115,6 @@ class ImageCommentsUIIntegrationTests: XCTestCase {
 		return (sut, loader)
 	}
 
-	func localized(_ key: String, file: StaticString = #filePath, line: UInt = #line) -> String {
-		let table = "ImageComments"
-		let bundle = Bundle(for: ImageCommentsPresenter.self)
-		let value = bundle.localizedString(forKey: key, value: nil, table: table)
-		if value == key {
-			XCTFail("Missing localized string for key: \(key) in table: \(table)", file: file, line: line)
-		}
-		return value
-	}
-
 	struct ExpectedCellContent {
 		let username: String
 		let message: String
